rm tests/metrics/diatomics/conftest.py
